third_party/WebKit/LayoutTests/shadow-dom/cascade-upgrade-from-v0-to-v1.html

<!DOCTYPE html>
<script src='../resources/testharness.js'></script>
<script src='../resources/testharnessreport.js'></script>
<style>
div { color: red; }
</style>
<div id='host'></div>
<div id='dummy'></div>
<script>
'use strict';

var root = host.createShadowRoot();
root.innerHTML = '<style>:host(#host) { color: green; }</style>';

test(() => {
  assert_equals(window.getComputedStyle(host).color, 'rgb(0, 128, 0)');
  // Trigger "V1" shadow cascading order.
  dummy.attachShadow({mode: 'open'});
  assert_equals(window.getComputedStyle(host).color, 'rgb(255, 0, 0)');
}, 'Upgrading V0 to V1 should cause style recalculation.');
</script>
